Fix test deployment issue - renamed test file to avoid Jest confusion (#501)
