+++ webui/src/environments/environment.testci.ts
@@ -1,0 +1,9 @@
+export const environment = {
+  name: "test",
+  production: true,
+  apiBaseUrl: 'https://test-webui.conformance-test-1.dcsa.org/',
+  useCognito: true,
+  cognitoUserPoolId: 'eu-north-1_VRXXJfTsv',
+  cognitoClientId: '4b2c2dt168j0tkf1ubg16n3g42',
+  siteTitle: 'DCSA Conformance Test',
+};
